2.25: Updating unit tests and classes for better coverage

+++ include/Data.hpp
@@ -30,7 +30,7 @@ class Data {
   Detect human_detector;
   cv::Mat frame;
   int getCamera(int mode);
-  int loadVideo(std::string filePath);
+  double loadVideo(std::string filePath);
   cv::Mat preProcessing(const cv::Mat &frame);
   cv::Mat videoPreProcessing(const cv::Mat &frame);
   ~Data();
